requirements.txt
aniso8601==9.0.1
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.0
click==8.0.3
cryptography==37.0.4
Flask==2.0.2
Flask-Login==0.6.2
flask-marshmallow==0.14.0
Flask-RESTful==0.3.9
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
idna==3.3
itsdangerous==2.0.1
Jinja2==3.0.3
jwt==1.3.1
MarkupSafe==2.0.1
marshmallow==3.17.0
marshmallow-sqlalchemy==0.28.1
packaging==21.3
pycparser==2.21
PyJWT==2.4.0
pyparsing==3.0.9
pytz==2022.2.1
requests==2.28.1
six==1.16.0
SQLAlchemy==1.4.27
urllib3==1.26.11
Werkzeug==2.0.2
